requirements.txt
MouseInfo==0.1.3
numpy==1.24.2
opencv-python-headless==4.7.0.72
opencv-python==4.7.0.72
pip==23.0.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
screeninfo==0.8.1
setuptools==65.5.1
termcolor==2.2.0
wheel==0.38.4
colorama~=0.4.6
Pillow~=9.5.0
pywin32~=306
